plainbox:testing_utils: fix TestCaseWithParameters on python3.4
This patch makes it possible for get_parameter_values() to return a
decorator. This seemed to have been silently supported on python3.3 as
the TestSuite class didn't make much use of the
TestCase.countTestCases() method but on python3.4 it was clearly broken.

This patch adds a fallback that enumerates all parameter values and
counts them. A simple regression test is added as well.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>