update test-in-vagrant.sh with relevant releases to be tested.
Will be testing:
- precise 12.04  (current LTS - supported until April 2017)
- trusty 14.04 (development version to be supported until April 2019)

We're ceasing per-merge testing of quantal 12.10 (supported until April
2014, but no longer used for any relevant projects) and raring 13.04
(EOL as of January 27th, 2013). Their definitions are still kept in the
Vagrantfile in case they are needed.

Note that Saucy will not be tested at the moment due to problems with
the vagrant cloud-images (http://pad.lv/1252872)

Signed-off-by: Daniel Manrique <roadmr@ubuntu.com>